web/requirements.txt
python-pptx==0.6.23
httpx==0.27.0